Added {% reverse %} & fixed formatting for signup
I added {% reverse %} as the action, but please not I did NOT test this on my site! I only tested it where you manually put in the newsletter name that you want the person to register for. I'm honestly a bit nervous about whether {% reverse %} will work or not, I'm not sure how it will know which newsletter without a view but maybe I don't understand the inner workings of django-newsletter well enough. 

I also fixed the formatting so that code blocks appear correctly.